Add useDeviceList hook
